chore: add missing error classes
