+++ src/account/memory.js
@@ -26,7 +26,7 @@ import AccountBase from './base'
 import { sign, isAddressValid, isValidKeypair } from '../utils/crypto'
 import { isHex } from '../utils/string'
 import { decode } from '../tx/builder/helpers'
-import { InvalidKeypairError, InvalidGAaddressError } from '../utils/error'
+import { InvalidKeypairError, InvalidGaAddressError } from '../utils/errors'
 
 const secrets = new WeakMap()
 
@@ -45,7 +45,7 @@ export default AccountBase.compose({
   init ({ keypair, gaId }) {
     this.isGa = !!gaId
     if (gaId) {
-      if (!isAddressValid(gaId)) throw new InvalidGAaddressError()
+      if (!isAddressValid(gaId)) throw new InvalidGaAddressError()
       secrets.set(this, { publicKey: gaId })
       return
     }
