+++ src/utils/aepp-wallet-communication/connection/browser-runtime.js
@@ -32,7 +32,7 @@ import {
   RpcConnectionError,
   AlreadyConnectedError,
   NoWalletConnectedError
-} from '../../error'
+} from '../../errors'
 
 /**
  * Disconnect
