--- src/utils/error.ts
@@ -1,653 +1,0 @@
-/**
- * Error module
- * @module @aeternity/aepp-sdk/es/utils/error
- * @example import { BytecodeMismatchError } from '@aeternity/aepp-sdk'
- */
-
-/**
- * aepp-sdk originated error
- */
-export class AeError extends Error {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'AeError'
-  }
-}
-
-export class CryptographyError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'CryptographyError'
-  }
-}
-
-export class WalletError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'WalletError'
-  }
-}
-
-export class AeppError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'AeppError'
-  }
-}
-
-export class SwaggerError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'SwaggerError'
-  }
-}
-
-/* Common error patterns */
-export class IllegalArgumentError extends CryptographyError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'IllegalArgumentError'
-  }
-}
-
-export class RequestTimedOutError extends AeError {
-  constructor (requestTime: number);
-  constructor (requestTime: number, currentHeight?: number, height?: number) {
-    if (currentHeight !== undefined && height !== undefined) {
-      super(`Giving up after ${requestTime}ms, current height: ${currentHeight}, desired height: ${height}`)
-    } else {
-      super(`Giving up after ${requestTime} ms`)
-    }
-    this.name = 'RequestTimedOutError'
-  }
-}
-
-export class TxTimedOutError extends AeError {
-  constructor (blocks: number, th: string) {
-    super(`Giving up after ${blocks} blocks mined, transaction hash: ${th}`)
-    this.name = 'TxTimedOutError'
-  }
-}
-
-export class TypeError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'TypeError'
-  }
-}
-
-export class MissingParamError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'MissingParamError'
-  }
-}
-
-export class InsufficientBalanceError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'InsufficientBalanceError'
-  }
-}
-
-export class UnsupportedProtocolError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'UnsupportedProtocolError'
-  }
-}
-
-export class InvalidNameError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'InvalidNameError'
-  }
-}
-
-export class InvalidHashError extends AeError {
-  constructor () {
-    super('Not a valid hash')
-    this.name = 'InvalidHashError'
-  }
-}
-
-export class UnknownHashClassError extends AeError {
-  constructor (klass: string) {
-    super(`Unknown class ${klass}`)
-    this.name = 'UnknownHashClassError'
-  }
-}
-
-export class InvalidDenominationError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'InvalidDenominationError'
-  }
-}
-
-export class NoSerializerFoundError extends AeError {
-  constructor () {
-    super('Byte serialization not supported')
-    this.name = 'NoSerializerFoundError'
-  }
-}
-
-export class UnsupportedPlatformError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'UnsupportedPlatformError'
-  }
-}
-
-export class NoBrowserFoundError extends AeError {
-  constructor () {
-    super('Browser is not detected')
-    this.name = 'NoBrowserFoundError'
-  }
-}
-
-/* cryptography errors */
-export class InvalidChecksumError extends CryptographyError {
-  constructor () {
-    super('Invalid checksum')
-    this.name = 'InvalidChecksumError'
-  }
-}
-
-export class MessageLimitError extends CryptographyError {
-  constructor () {
-    super('message too long')
-    this.name = 'MessageLimitError'
-  }
-}
-
-export class InvalidDerivationPathError extends CryptographyError {
-  constructor () {
-    super('Invalid path')
-    this.name = 'InvalidDerivationPathError'
-  }
-}
-
-export class NotHardenedSegmentError extends CryptographyError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'NotHardenedSegmentError'
-  }
-}
-
-export class UnsupportedChildIndexError extends CryptographyError {
-  constructor (index: string) {
-    super(`Child index #${index} is not supported`)
-    this.name = 'UnsupportedChildIndexError'
-  }
-}
-
-export class InvalidMnemonicError extends CryptographyError {
-  constructor () {
-    super('Invalid mnemonic')
-    this.name = 'InvalidMnemonicError'
-  }
-}
-
-export class MerkleTreeHashMismatchError extends CryptographyError {
-  constructor () {
-    super('Node hash is not equal to provided one')
-    this.name = 'MerkleTreeHashMismatchError'
-  }
-}
-
-export class MissingNodeInTreeError extends CryptographyError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'MissingNodeInTreeError'
-  }
-}
-
-export class UnknownPathNibbleError extends CryptographyError {
-  constructor (nibble: number) {
-    super(`Unknown path nibble: ${nibble}`)
-    this.name = 'UnknownPathNibbleError'
-  }
-}
-
-export class UnknownNodeLengthError extends CryptographyError {
-  constructor (nodeLength: number) {
-    super(`Unknown node length: ${nodeLength}`)
-    this.name = 'UnknownNodeLengthError'
-  }
-}
-
-export class InvalidKeyError extends CryptographyError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'InvalidKeyError'
-  }
-}
-
-export class NoSuchAlgorithmError extends CryptographyError {
-  constructor (algo: string) {
-    super(algo + ' is not available')
-    this.name = 'NoSuchAlgorithmError'
-  }
-}
-
-export class UnsupportedKdfError extends CryptographyError {
-  constructor () {
-    super('Unsupported kdf type')
-    this.name = 'UnsupportedKdfError'
-  }
-}
-
-export class InvalidPasswordError extends CryptographyError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'InvalidPasswordError'
-  }
-}
-/* keypair an account related errors */
-export class InvalidKeypairError extends IllegalArgumentError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'InvalidKeypairError'
-  }
-}
-
-export class UnavailableAccountError extends AeError {
-  constructor (address: string) {
-    super(`Account for ${address} not available`)
-    this.name = 'UnavailableAccountError'
-  }
-}
-
-/* errors originating from AENS operation */
-export class InsufficientNameFeeError extends AeError {
-  constructor (nameFee: number, minNameFee: number) {
-    super(`the provided fee ${nameFee} is not enough to execute the claim, required: ${minNameFee}`)
-    this.name = 'InsufficientNameFeeError'
-  }
-}
-
-export class InvalidAensNameError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'InvalidAensNameError'
-  }
-}
-
-export class AensNameNotFoundError extends AeError {
-  constructor (nameOrId: string) {
-    super(`Name not found: ${nameOrId}`)
-    this.name = 'AensNameNotFoundError'
-  }
-}
-
-export class AensPointerContextError extends AeError {
-  constructor (nameOrId: string, prefix: string) {
-    super(`Name ${nameOrId} don't have pointers for ${prefix}`)
-    this.name = 'AensPointerContextError'
-  }
-}
-
-/* transaction related errors */
-export class InvalidTxError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'InvalidTxError'
-  }
-}
-
-export class TxNotInChainError extends AeError {
-  constructor (txHash: string) {
-    super(`Transaction ${txHash} is removed from chain`)
-    this.name = 'TxNotInChainError'
-  }
-}
-
-export class DryRunError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'DryRunError'
-  }
-}
-
-export class UnsupportedVMversionError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'UnsupportedVMversionError'
-  }
-}
-
-export class UnsupportedABIversionError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'UnsupportedABIversionError'
-  }
-}
-
-export class UnknownTxError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'UnknownTxError'
-  }
-}
-
-export class UnsignedTxError extends AeError {
-  constructor () {
-    super('Signature not found, transaction is not signed')
-    this.name = 'UnsignedTxError'
-  }
-}
-
-export class InvalidSignatureError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'UnsignedTxError'
-  }
-}
-
-export class SchemaNotFoundError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'SchemaNotFoundError'
-  }
-}
-
-export class InvalidTxParamsError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'InvalidTxParamsError'
-  }
-}
-
-export class PrefixMismatchError extends AeError {
-  constructor (prefix: string, requiredPrefix: string) {
-    super(`Encoded string have a wrong type: ${prefix} (expected: ${requiredPrefix})`)
-    this.name = 'PrefixMismatchError'
-  }
-}
-
-export class DecodeError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'DecodeError'
-  }
-}
-
-export class TagNotFoundError extends AeError {
-  constructor (prefix: string) {
-    super(`Id tag for prefix ${prefix} not found.`)
-    this.name = 'DecodeError'
-  }
-}
-
-export class PrefixNotFoundError extends AeError {
-  constructor (tag: string) {
-    super(`Prefix for id-tag ${tag} not found.`)
-    this.name = 'PrefixNotFoundError'
-  }
-}
-
-export class IllegalBidFeeError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'IllegalBidFeeError'
-  }
-}
-
-/* compiler issued errors */
-export class UnsupportedCompilerError extends AeError {
-  constructor (compilerVersion: string, COMPILER_GE_VERSION: string, COMPILER_LT_VERSION: string) {
-    super(`Unsupported compiler version ${compilerVersion}. ` +
-    `Supported: >= ${COMPILER_GE_VERSION} < ${COMPILER_LT_VERSION}`)
-    this.name = 'UnsupportedCompilerError'
-  }
-}
-
-export class UnavailableCompilerError extends AeError {
-  constructor () {
-    super('Compiler is not ready')
-    this.name = 'UnavailableCompilerError'
-  }
-}
-
-/* errors from option validations */
-export class InvalidAuthDataError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'InvalidAuthDataError'
-  }
-}
-
-/* contract invocation errors */
-export class MissingContractDefError extends AeError {
-  constructor () {
-    super('Either ACI or source code is required')
-    this.name = 'MissingContractDefError'
-  }
-}
-
-export class MissingContractAddressError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'MissingContractAddressError'
-  }
-}
-
-export class NoSuchContractError extends AeError {
-  constructor (notExistingContractAddress: string) {
-    super(`Contract with address ${notExistingContractAddress} not found on-chain`)
-    this.name = 'NoSuchContractError'
-  }
-}
-
-export class InactiveContractError extends AeError {
-  constructor (contractAddress: string) {
-    super(`Contract with address ${contractAddress} not active`)
-    this.name = 'InactiveContractError'
-  }
-}
-
-export class BytecodeMismatchError extends AeError {
-  constructor (source: string) {
-    super(`Contract ${source} do not correspond to the bytecode deployed on the chain`)
-    this.name = 'BytecodeMismatchError'
-  }
-}
-
-export class UnknownCallReturnTypeError extends AeError {
-  constructor (returnType: any) {
-    super(`Unknown returnType: ${String(returnType)}`)
-    this.name = 'UnknownCallReturnTypeError'
-  }
-}
-
-export class NodeInvocationError extends AeError {
-  constructor (msg?: string) {
-    msg = msg ?? ''
-    super(`Invocation failed${msg === '' ? '' : `: "${msg}"`}`)
-    this.name = 'NodeInvocationError'
-  }
-}
-
-export class DuplicateContractError extends AeError {
-  constructor () {
-    super('Contract already deployed')
-    this.name = 'DuplicateContractError'
-  }
-}
-
-export class MissingFunctionNameError extends AeError {
-  constructor () {
-    super('Function name is required')
-    this.name = 'MissingFunctionNameError'
-  }
-}
-
-export class InvalidMethodInvocationError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'InvalidMethodInvocationError'
-  }
-}
-
-export class NotPayableFunctionError extends AeError {
-  constructor (amount: string, fn: string) {
-    super(
-      `You try to pay "${amount}" to function "${fn}" which is not payable. ` +
-      'Only payable function can accept tokens')
-    this.name = 'NotPayableFunctionError'
-  }
-}
-
-export class InvalidSchemaError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'InvalidSchemaError'
-  }
-}
-
-export class NoSuchContractFunctionError extends AeError {
-  constructor (name: string) {
-    super(`Function ${name} doesn't exist in contract`)
-    this.name = 'NoSuchContractFunctionError'
-  }
-}
-
-/* Node communication errors  */
-export class UnsupportedNodeError extends AeError {
-  constructor (nodeVersion: string, NODE_GE_VERSION: string, NODE_LT_VERSION: string) {
-    super(
-      `Unsupported node version ${nodeVersion}. ` +
-      `Supported: >= ${NODE_GE_VERSION} < ${NODE_LT_VERSION}`)
-    this.name = 'UnsupportedNodeError'
-  }
-}
-
-/* Channel originated errors */
-export class ChannelInitializationError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'ChannelInitializationError'
-  }
-}
-
-export class UnknownChannelStateError extends AeError {
-  constructor () {
-    super('State Channels FSM entered unknown state')
-    this.name = 'UnknownChannelStateError'
-  }
-}
-
-export class ChannelPingTimedOutError extends AeError {
-  constructor () {
-    super('Server pong timed out')
-    this.name = 'ChannelPingTimedOutError'
-  }
-}
-
-export class ChannelCallError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'ChannelCallError'
-  }
-}
-
-export class UnexpectedChannelMessageError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'UnexpectedChannelMessageError'
-  }
-}
-
-export class ChannelConnectionError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'ChannelConnectionError'
-  }
-}
-
-/* Node errors */
-export class NodeNotFoundError extends AeError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'NodeNotFoundError'
-  }
-}
-
-export class DuplicateNodeError extends AeError {
-  constructor (name: string) {
-    super(`Node with name ${name} already exist`)
-    this.name = 'DuplicateNodeError'
-  }
-}
-
-export class DisconnectedError extends AeError {
-  constructor () {
-    super('Can not get node info. Node is not connected')
-    this.name = 'DisconnectedError'
-  }
-}
-
-/* Wallet & Aepp related Errors */
-export class RpcConnectionError extends WalletError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'RpcConnectionError'
-  }
-}
-
-export class AlreadyConnectedError extends WalletError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'AlreadyConnectedError'
-  }
-}
-
-export class NoWalletConnectedError extends WalletError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'NoWalletConnectedError'
-  }
-}
-
-export class MessageDirectionError extends WalletError {
-  constructor (msg: string) {
-    super(msg)
-    this.name = 'MessageDirectionError'
-  }
-}
-
-export class UnsubscribedAccountError extends AeppError {
-  constructor () {
-    super('You are not subscribed for an account.')
-    this.name = 'UnsubscribedAccountError'
-  }
-}
-
-export class UnAuthorizedAccountError extends AeppError {
-  constructor (onAccount: string) {
-    super(`You do not have access to account ${onAccount}`)
-    this.name = 'UnAuthorizedAccountError'
-  }
-}
-
-export class UnknownRpcClientError extends AeppError {
-  constructor (id: string) {
-    super(`RpcClient with id ${id} do not exist`)
-    this.name = 'UnknownRpcClientError'
-  }
-}
-
-export class InvalidRpcMessage extends AeppError {
-  constructor (msg: string) {
-    super(`Received invalid message: ${msg}`)
-    this.name = 'InvalidRpcMessage'
-  }
-}
-
-export class MissingCallbackError extends AeppError {
-  constructor (id: string) {
-    super(`Can't find callback for this messageId ${id}`)
-    this.name = 'MissingCallbackError'
-  }
-}
