+++ test/integration/node.js
@@ -20,7 +20,7 @@ import { url, internalUrl, ignoreVersion } from './'
 import { describe, it, before } from 'mocha'
 import { expect } from 'chai'
 import NodePool from '../../src/node-pool'
-import { MissingParamError, NodeNotFoundError, TypeError } from '../../src/utils/error'
+import { MissingParamError, NodeNotFoundError, TypeError } from '../../src/utils/errors'
 
 describe('Node client', function () {
   let node
