+++ test/unit/mptree.ts
@@ -22,7 +22,7 @@ import { decode as rlpDecode } from 'rlp'
 import type { Input } from 'rlp'
 import MPTree from '../../src/utils/mptree'
 import type { MPTreeBinary } from '../../src/utils/mptree'
-import { MerkleTreeHashMismatchError, MissingNodeInTreeError, UnknownNodeLengthError } from '../../src/utils/error'
+import { MerkleTreeHashMismatchError, MissingNodeInTreeError, UnknownNodeLengthError } from '../../src/utils/errors'
 
 const hexToTreeBinary = (hex: string): MPTreeBinary => {
   return rlpDecode(Buffer.from(hex, 'hex') as Input) as MPTreeBinary
